Add files via upload
The current mania over Artificial Intelligence (AI) reminded me of a program I wrote for the original Atari 800 in 1984, for a post-graduate private computer school class.  I've left the manual describing the program completely as is, except for redacting my old address on the cover page.  Apologies for the yellowing pages, which are 39 years old now.

The slight skewing from the scanner didn't lose any characters in the coding section and you can still type in the full code if you desire.  I submit this as open-source and as a learning tool, to see how a basic - also BASIC - AI program was written in the very early days, almost before the IBM PC came out.

It is based upon a very simple AI program in Compute! magazine - just 23 lines, or <1/10th the size of my greatly enhanced program.

The program was originally designed for the Atari 800 but ought to run on the Atari 1450xld with speech options.  The 1450xld was never officially released before Atari declared bankruptcy, but I was able to find the BASIC code that would have worked for that machine, and included it in this late stage Atari program, completed by May, 1984.

Superlearner also has an option to print what it has learned to a printer, or save the output to a disk or cassette.

The Atari had no access to the internet, which, in any case, did not exist in 1984, so there was no option for a BASIC AI program except for the user to manually enter facts and non-facts about subjects for the AI program to mull over.  Menu options make this a simple matter.  Everything is menu-driven, including the sample data already built into the program.

I lost the function of my Atari 410 cassette recorder a long time ago and although I have a working Atari 810 disk drive, I have not had the patience to type in the program again.  If there is sufficient interest in the program I may do so again, and provide some sample output from the program.

There are some similarities to today's AI but of course, the level of complexity and capacity is millions of orders of magnitude greater today.  I present this as a simple example of AI and of historical background to today's AI apps.

Enjoy!